moving main database functions to the save data file
